Reporting feature code checkin as per review comments
