+++ DCode.Web/Controllers/ReportingBaseController.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 using DCode.Web.Security;
 
 namespace DCode.Web.Controllers
