+++ DCode.Web/Views/Reporting/GetFilteredUsers.cshtml
@@ -7,11 +7,8 @@
 
 <h2>Filtered Users</h2>
 
-<p>
-    @*@Html.ActionLink("Create New", "Create")*@
-</p>
 <form action="GetFilteredUsers" method="post">
-    <table class="table table-bordered table-hover table-responsive table-striped" >
+    <table class="table table-bordered table-hover table-responsive table-striped">
         <tr>
             <th>
                 @Html.DropDownList("ddl_Desg", new List<SelectListItem>
@@ -32,13 +29,10 @@
             <th>
                 <input type="submit" value="ApplyFilter" id="btnSubmit" />
             </th>
-            <th>
-                @Html.ActionLink("Export To Excel", "ExportToExcel", "Users")
-            </th>
         </tr>
     </table>
 </form>
-<table  class="table table-bordered table-hover table-responsive table-striped">
+<table class="table table-bordered table-hover table-responsive table-striped">
     <tr>
         <th>
             @Html.DisplayNameFor(model => model.FIRST_NAME)
@@ -85,56 +79,51 @@
         <th></th>
     </tr>
 
-@foreach (var item in Model) {
-    <tr>
-        <td>
-            @Html.DisplayFor(modelItem => item.FIRST_NAME)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.LAST_NAME)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.DESIGNATION)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.EMAIL_ID)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.PROJECT_NAME)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.PROJECT_CODE)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.PROJECT_MANAGER_NAME)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.MANAGER_EMAIL_ID)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.STATUS)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.STATUS_DATE)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.CREATED_BY)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.CREATED_ON)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.UPDATED_BY)
-        </td>
-        <td>
-            @Html.DisplayFor(modelItem => item.UPDATED_ON)
-        </td>
-        <td>
-            @*@Html.ActionLink("Edit", "Edit", new { id=item.ID }) |
-            @Html.ActionLink("Details", "Details", new { id=item.ID }) |
-            @Html.ActionLink("Delete", "Delete", new { id=item.ID })*@
-        </td>
-    </tr>
-}
-
+    @foreach (var item in Model)
+    {
+        <tr>
+            <td>
+                @Html.DisplayFor(modelItem => item.FIRST_NAME)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.LAST_NAME)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.DESIGNATION)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.EMAIL_ID)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.PROJECT_NAME)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.PROJECT_CODE)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.PROJECT_MANAGER_NAME)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.MANAGER_EMAIL_ID)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.STATUS)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.STATUS_DATE)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.CREATED_BY)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.CREATED_ON)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.UPDATED_BY)
+            </td>
+            <td>
+                @Html.DisplayFor(modelItem => item.UPDATED_ON)
+            </td>
+        </tr>
+    }
 </table>
