+++ DCode.Web/Views/Reporting/TopContributors.cshtml
@@ -4,22 +4,7 @@
     ViewBag.Title = "Top Contributors List";
     Layout = "~/Views/Shared/_LayoutReporting.cshtml";
 }
-
 <h2>Top Contributors List</h2>
-
-<form method="post" action="">
-
-    <table>
-        <tr>
-            <th>
-                @Html.ActionLink("Export To Excel", "ExportToExcel", "Reporting")
-            </th>
-        </tr>
-    </table>
-
-</form>
-
-
 <table class="table table-bordered table-responsive table-hover">
     <tr>
         <th>
@@ -31,9 +16,7 @@
         <th>
             @Html.DisplayNameFor(model => model.Count)
         </th>
-
     </tr>
-
     @foreach (var item in Model)
     {
         <tr>
@@ -48,7 +31,6 @@
             </td>
         </tr>
     }
-
 </table>
