+++ DCode.Web/Views/ReportingBase/Base.cshtml
@@ -9,21 +9,11 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>@ViewBag.Title - Reporting Page</title>
-    @Styles.Render("~/Content")
-    <link href="~/Content/bootstrap.css" rel="stylesheet" />
-    @* @Scripts.Render("~/bundles/")*@
 </head>
 <body>
     <div class="navbar navbar-inverse navbar-fixed-top">
-      Reporting content here   
+        Reporting content here
     </div>
-
-  
-
-
-    @Scripts.Render("~/bundles/jquery")
-    @Scripts.Render("~/bundles/bootstrap")
-
 </body>
 </html>
